requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.10
Pillow==8.2.0
pipreqs==0.4.10
psycopg2==2.8.6
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
yarg==0.1.9
